--Major bug fix from last PR--
---MAJOR BUG---
From last branch marge (bread-crumbs-in-404-and-error) When entering an exact url of a single mattress would throw an error sometimes due to using  `mattress.brandName' graphql data before the loading state of the component. Which is used to pass props to the error component. Fix was to use 'match.path' from react router instead.